Add a new presubmit environment for bzlmod (#2309)
Add an environment which builds without a workspace file.

This will allow us to verify that targets such as cargo_bazel can build
with bzlmod. This is mainly useful to ensure that your dependencies are
specified in MODULE.bazel, rather than WORKSPACE. At the moment, I only
specify a single target, to verify that it's working. As we start
supporting additional targets, entries will be added here to prevent
regressions.