Adjust is_exec_configuration heuristic (#2349)
Instead of looking for "-exec-" look for "-exec" since that portion may
be the last component of the output path mneumonic.

Ultimately fixes issues with --experimental_platforms_in_output_dir that
had sufficiently improved calculations to remove the need for an ST-hash
and thus surfaced this issue.

Ideally, in the long term, this heuristic would not be used and whether
an artifact is in the exec configuration would be determined by tracking
the provenance of that artifact. However, this would likely require a
more significant rework.