Add script to revendor all crates (#2459)
It seems that vendored crates commonly get out of sync. Re-vendoring
seems to currently be a quite manual process, and finding everything to
re-vendor is a hassle.

This adds a bash script, `vendor.sh`, that finds all crate vendor rules
in all workspaces in this repository and runs them. This significantly
improves the workflow for updating these.

This also adds `.bazelrc` files to disable bzlmod to non-bzlmod
workspaces, since otherwise bazel 7 create a MODULE.bazel and
MODULE.bazel.lock when re-vendoring.

This also re-vendors all crates (in a separate commit).