fix rustdoc: omit link flags for rustdoc (#2467)
In our project, rust_doc fails because it is generating `-l` arguments
to
rustdoc, and rustdoc does not support `-l` arguments.

In my understanding, these arguments are only relevant for calling the
linker at some point, and it won't be called for rustdoc anyway.

With this PR, our project can build with rust_doc.

---------

Co-authored-by: Krasimir Georgiev <krasimir@google.com>