Fix (+) In vendored crates semver metadata (#2579)
- Resolves https://github.com/bazelbuild/rules_rust/issues/2524
- Revendored examples deps.
- Semver Metadata can break vendored dependencies, by cargo vendoring
(foo-1.2.3+1.2.3) for example. The bazel labels will look for
`foo-1.2.3-1.2.3` which does not work.
- Root of the issue is, rules rust uses cargo to vendor the
dependencies, which takes the semver + metadata, and generates the file
path based off that (hence the + for the dependency in the issue)
- Tested by taking that dependency and make sure it loaded using the
`vendor_local_manifests` example
- Did not include that dependency, because it would introduce a sys dep
in CI
- Opted to do it within the scope of the vendor cli. When trying to
change the version in the `render_build_files` call, it fixes the
version in terms of the `outputs` and `files` declared in the vendor cli
file, but then we do not have a handle to the generated vendor files
from cargo, meaning we cant move the contents over near as easily. I
also didn't want to break anything downstream that may rely on that
info.