Bzlmod-aware runfiles library (#2566)
This implements the repo_mapping capabilities and provides a new
runfiles API. The new API can be accessed explicitly as
`runfiles::Runfiles::rlocation_from` or with the `runfiles::rlocation!`
macro, which adds compile-time support for correctly embedding the
external repo. This is a purely new API, existing usage continues to
work, although we mark it deprecated because it's not fully correct. We
can remove it at some point in the future.

This PR also transitions in-repo examples/tests to using it, in case
anyone copies them.

---------

Co-authored-by: scentini <rosica@google.com>