Allow overriding crate universe packages to local packages (#2674)
Allow external crates to be aliased to local targets

In some cases it may be desirable to alias a public crate to a locally
define rust_library target, even for dependencies of crates that are not
being overriden.
Support this use case by swapping out `alias` for `rust_library` when
a user overrides the target using an annotation.

---------

Co-authored-by: Daniel Wagner-Hall <dwagnerhall@apple.com>