Depend on a copy of rustfmt for the target (#2685)
The toolchain_files() and current_rust_toolchain() rules return files
for the current 'exec' platform. This is perfecly reasonable if you want
to use these tools as part of ctx.actions.run(). But if you want to use
these as part of 'data = []' (runfiles), they must be built for the
target.

Fixes: https://github.com/bazelbuild/rules_rust/issues/2684