Add support for override_target when using bzlmod (#2683)
The ability to override crate targets when using crate_universe was
added here: https://github.com/bazelbuild/rules_rust/pull/2674

However, this change did not expose that functionality when using
bzlmod.

This change adds that functionality in. Because of the limited set of
options we have for dictionaries in tag classes, I had to split out
"override_targets" into four different options, each taking a Label.

I have added in an example based on the example given in #2674 , but
using bzlmod instead. I have also tested that example and found it to be
working.