Apply extra_rustc_flags list to all toolchains (#2693)
Fixes #2692 

This fixes `rust_register_toolchains` to correctly call
`rust_toolchain_repository` with a list of `extra_rustc_flags`.

Previously, the `extra_rustc_flags` argument of
`rust_register_toolchains` was assumed (when set) to be a dict of
toolchain triple to list. With this change, a `list` parameter is used
for all toolchain triples.

Co-authored-by: UebelAndre <github@uebelandre.com>